Implemented a timer in ParameterFunction do reduce the number of updates I am sending to LR, as it simply couldn't handle it and was lagging behind on many simultaneous changes. Currently clamped the push rate to 60/s for each parameter function

+++ Source/LrControl/Functions/EnableFunctionGroupFunction.cs
@@ -17,7 +17,7 @@ namespace micdah.LrControl.Functions
 
         protected override void ControllerChanged(int controllerValue)
         {
-            if (controllerValue != ControllerRange.Maximum) return;
+            if (controllerValue != (int)ControllerRange.Maximum) return;
 
             if (!_functionGroup.Enabled)
             {
